+++ @xen-orchestra/lite/CHANGELOG.md
@@ -7,6 +7,7 @@
 - [Icons] Update FontAwesome, clean unused libraries (PR [#10365](https://github.com/vatesfr/xen-orchestra/pull/10365))
 - [Host] Add possibility to shut down and start a host (PR [#10314](https://github.com/vatesfr/xen-orchestra/pull/10314))
 - Use `MenuItem`'s `accent` prop to color menu actions instead of custom CSS classes (PR [#10348](https://github.com/vatesfr/xen-orchestra/pull/10348))
+- [Pool/Network] Add the possibility to delete host internal networks
 
 ## **0.25.0** (2026-08-31)